Remove most of `qiskit.test`
This removes large tracts of `qiskit.test` that Qiskit itself is no
longer using, or inlines a couple of components that were either
exceptionally simple or used only once.  A few tests that made heavier
use of the removed features have been removed rather than replacing the
functionality, since those had not actually been running due to
dependencies on IBMQ that have not been fulfilled for a long time.

This prepares for a more complete removal of `qiskit.test`, if we choose
to, but such a change raises the possibility of rather more merge
conflicts (due to the need to touch every test file), so is best done
separately.
