+++ releasenotes/notes/remove-qiskit-test-8f3f43ed2a185f2f.yaml
@@ -1,0 +1,7 @@
+---
+upgrade:
+  - |
+    The ``qiskit.test`` module is no longer a public module.  This was never intended to be public,
+    nor used outside of Qiskit's own test suite.  All functionality was specific to Qiskit and no
+    alternative is provided; if you needed similar functionality, you should include it in your own
+    test harnesses.
